requirements.txt
astronomer-cosmos>=1.0.2
dbt-core==1.5.10
dbt-postgres==1.5.10